+++ src/main/java/roomescape/controller/ReservationController.java
@@ -1,4 +1,4 @@
-package roomescape.controller.reservation;
+package roomescape.controller;
 
 import java.util.List;
 import org.springframework.http.ResponseEntity;
@@ -9,9 +9,9 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.controller.reservation.request.ReservationRequest;
-import roomescape.controller.reservation.response.ReservationResponse;
 import roomescape.service.ReservationService;
+import roomescape.service.dto.reservation.request.ReservationRequest;
+import roomescape.service.dto.reservation.response.ReservationResponse;
 
 @RequestMapping("/reservations")
 @RestController
