+++ src/main/java/roomescape/service/ReservationTimeService.java
@@ -2,10 +2,10 @@ package roomescape.service;
 
 import java.util.List;
 import org.springframework.stereotype.Service;
-import roomescape.controller.reservationtime.request.ReservationTimeRequest;
-import roomescape.controller.reservationtime.response.ReservationTimeResponse;
 import roomescape.model.ReservationTime;
 import roomescape.repository.ReservationTimeDao;
+import roomescape.service.dto.reservationtime.request.ReservationTimeRequest;
+import roomescape.service.dto.reservationtime.response.ReservationTimeResponse;
 
 @Service
 public class ReservationTimeService {
