+++ src/main/java/roomescape/service/dto/reservationtime/response/ReservationTimeResponse.java
@@ -1,4 +1,4 @@
-package roomescape.controller.reservationtime.response;
+package roomescape.service.dto.reservationtime.response;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import java.time.LocalTime;
